requirements.txt
langchain
langchain_groq
streamlit
speechrecognition
pyttsx3
pyaudio
python-dotenv
